prepare environment
The stack of my choice is:
- pure Node.js as runtime (including watch mode and TS support)
- Express.js as API framework
- Vitest for testing (I know Node.js has built-in testing, but I want to try Vitest)
- Prettier for code formatting because it is annoying to do it myself

No eslint, no build tool for now. I will start with server side, because that has to be functional for the frontend anyway AND it can be done in TDD.

Also note there is no mention of Docker for now - that is on purpose. I'd develop natively anyway, because Docker Compose is just sooo slow. Production docker image will come later.