implement `POST /create` CSV adapter
The idea of adapters is that you can easily switch between various storage layers. The one I implemented is a CSV file. Whole app was restructured to accept an adapter as a parameter. This is very handy for testing, because we can replace prod-ready adapter with a test version that we can easily "look" into - no mocking needed.

A note to CsvTestAdapter: it doesn't do any CSV magic, because I don't want to test external libraries - I just believe they work.