dockerize server
Final step for server API is dockerizing it so it can easily be deployed anywhere. Even your local machine without dependencies! Anyway, I had to do few adjustments to imports because Node.js typescript support is very limited - it is unable to determine what imports are types, I have to mark them manually.

Also, I added a new .gitignore entry for the CSV file that is generated by the server. It is not needed in the repository.