fix csv adapter bugs
While testing frontend, I can see so many bugs in the csv adapter! I think it wasn't a good idea to call test adapter an alternative to CSV adapter. I think `CsvTestAdapter` is actually `MemoryAdapter` - a great help for test speed, but `CsvAdapter` needs specialized tests that ensure CSV file is updated as expected. All of that will come in a next commit.